+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add actor ID to "signer not found in cache" error message.
 - Make single HTTP request when searching for profile/post by URL.
 - Reduced compilation time by disabling unused Actix features.
+- Don't panic if compatible ID is stored in database.
 
 ## [4.19.0] - 2026-02-27